sql: move BackupRestoreTestingKnobs to exec_util_backup.go
`BackupRestoreTestingKnobs` has been moved to a new file,
`exec_util_backup.go` and `CODEOWNERS` has been updated.

Release note: None
